backend/requirements.txt
fastapi
uvicorn
pydantic
slowapi
pinecone
langgraph
langchain
langchain-core
langchain-huggingface
langchain-openai
langchain-text-splitters
sentence-transformers
pdfplumber
PyPDF2
fpdf
requests
python-dotenv
openai
rank_bm25
